+++ src/fdm_edl/solver/__init__.py
@@ -2,9 +2,11 @@
 """Nonlinear root-finding solvers."""
 
 from .base import BaseSolver
-from .naive_newton import NewtonSolver
+from .newton import NewtonSolver
+from .tmp import BiCGStabSolver
 
 __all__ = [
     "BaseSolver",
     "NewtonSolver",
+    "BiCGStabSolver",
 ]